refactor packages
move StringParser class to common folder.
delete the parser folder which only had StringParser class inside.
