+++ src/test/java/seedu/StringParserTest.java
@@ -2,7 +2,7 @@ package seedu;
 
 import org.junit.jupiter.api.Test;
 import seedu.typists.common.exception.InvalidStringInputException;
-import seedu.typists.parser.StringParser;
+import seedu.typists.common.StringParser;
 
 import java.util.ArrayList;
